* libtool: version up to 1.5.26   autoconf: version up to 2.63   (for win32 compilation.)
* Separated mkf table from mkf as dynamic shared library(LIBDIR/mkf/libmkf_*.so).
  (Added configure option --enable(disable)-dl-table)

* Added scrollbar support in win32.
  (Added x_simple_sb_view_win32.c)

* x_window_win32.c: Apply min_{width|height} and {width|height}_inc
                    when window resized.

* *_color_*.[ch]: u_short changed to u_int8_t.

* etc/mlterm/color: Supported rgb:RR/GG/BB format as value.

* mlterm/ml_color.[ch]: Added ml_color_parse_rgb_name() function.

+++ mkf/lib/mkf_ucs4_johab.c
@@ -9,37 +9,37 @@
 
 /* 32 = 2^5 */
 
-static int  johab_first_to_linear[32] =
+static int8_t  johab_first_to_linear[32] =
 {
 	0 , 0 , 1 , 2 , 3 , 4 , 5 , 6 , 7 , 8 , 9 , 10 , 11 , 12 , 13 , 14 , 15 , 16 ,
 	17 , 18 , 19 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0
 } ;
 
-static int  johab_middle_to_linear[32] =
+static int8_t  johab_middle_to_linear[32] =
 {
 	0 , 0 , 0 , 1 , 2 , 3 , 4 , 5 , 0 , 0 , 6 , 7 , 8 , 9 , 10 , 11 ,
 	0 , 0 , 12 , 13 , 14 , 15 , 16 , 17 , 0 , 0 , 18 , 19 , 20 , 21 , 0 , 0
 } ;
 
-static int  johab_last_to_linear[32] =
+static int8_t  johab_last_to_linear[32] =
 {
 	0 , 1 , 2 , 3 , 4 , 5 , 6 , 7 , 8 , 9 , 10 , 11 , 12 , 13 , 14 , 15 ,
 	16 , 17 , 0 , 18 , 19 , 20 , 21 , 22 , 23 , 24 , 25 , 26 , 27 , 28 , 0 , 0
 } ;
 
-static int  linear_to_johab_first[32] =
+static int8_t  linear_to_johab_first[32] =
 {
 	2 , 3 , 4 , 5 , 6 , 7 , 8 , 9 , 10 , 11 , 12 , 13 , 14 , 15 , 16 , 17 ,
 	18 , 19 , 20 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0
 } ;
 
-static int  linear_to_johab_middle[32] =
+static int8_t  linear_to_johab_middle[32] =
 {
 	3 , 4 , 5 , 6 , 7 , 10 , 11 , 12 , 13 , 14 , 15 , 18 , 19 , 20 , 21 , 22 ,
 	23 , 26 , 27 , 28 , 29 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0
 } ;
 
-static int  linear_to_johab_last[32] =
+static int8_t  linear_to_johab_last[32] =
 {
 	1 , 2 , 3 , 4 , 5 , 6 , 7 , 8 , 9 , 10 , 11 , 12 , 13 , 14 , 15 , 16 ,
 	17 , 19 , 20 , 21 , 22 , 23 , 24 , 25 , 26 , 27 , 28 , 29 , 0 , 0 , 0 , 0
